Break out graphql_queries from helpers
